Give the network parsers a module of their own
_point.py, _track.py and _vertical.py each own their parse function;
network was the only geometry without one, so its two parsers sat in
_point.py. The edge is one way -- _network -> _point -> _coords -- and
timeseries/__init__ re-exports the same names, so no caller changes.
