+++ src/modelskill/timeseries/__init__.py
@@ -1,6 +1,6 @@
 from ._timeseries import TimeSeries
-from ._point import (
-    _parse_xyz_point_input,
+from ._point import _parse_xyz_point_input
+from ._network import (
     _parse_network_node_input,
     _parse_network_breakpoint_input,
 )
